refactor: organize codebase feature wise
