+++ apps/web/src/features/anime/components/add-to-list-dialog.tsx
@@ -26,9 +26,9 @@ import {
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { cn } from "@/lib/utils";
-import { LIBRARY_STATUSES, type LibraryEntryWithAnime } from "@/features/lists/lib/options";
-import { type LogAnimeData } from "@/features/lists/lib/options";
-import { useLogAnime, useRemoveFromLibrary } from "@/features/lists/lib/hooks";
+import { LIBRARY_STATUSES, type LibraryEntryWithAnime } from "@/features/library/lib/options";
+import { type LogAnimeData } from "@/features/library/lib/options";
+import { useLogAnime, useRemoveFromLibrary } from "@/features/library/lib/hooks";
 
 interface AddToListDialogProps {
   anime: LogAnimeData["anime"] | null;
