+++ apps/web/src/features/anime/lib/prefetch.server.ts
@@ -3,7 +3,7 @@ import "server-only";
 import type { QueryClient } from "@tanstack/react-query";
 
 import { animeQueries } from "@/features/anime/lib/options";
-import { libraryQueries } from "@/features/lists/lib/options";
+import { libraryQueries } from "@/features/library/lib/options";
 
 export async function prefetchAnimeHome(
   queryClient: QueryClient,
