+++ apps/web/src/features/users/components/user-card.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { Card, CardContent } from "@/components/ui/card";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { FollowButton } from "./follow-button";
 import type { UserWithProfile } from "../lib/options";
