+++ apps/web/src/features/users/components/user-lists-public.tsx
@@ -6,8 +6,8 @@ import type { LibraryStatus } from "@anilog/db/schema/anilog";
 import { useUserLists } from "../lib/hooks";
 import { Button } from "@/components/ui/button";
 import { AnimeCard } from "@/features/anime/components/anime-card";
-import { LIBRARY_STATUSES } from "@/features/lists/lib/options";
-import { AnimeStackPreview } from "@/features/lists/components/anime-stack-preview";
+import { LIBRARY_STATUSES } from "@/features/library/lib/options";
+import { AnimeStackPreview } from "@/features/library/components/anime-stack-preview";
 
 interface UserListsPublicProps {
   userId: string;
