+++ packages/api/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./services/anime.service";
+export * from "./services/admin.service";
 export * from "./services/library.service";
 export * from "./services/user.service";
 export * from "./errors/api-error";
