+++ packages/api/src/services/user.service.ts
@@ -1,7 +1,7 @@
 import { db } from "@anilog/db";
 import { userFollow, userProfile, userAnime, anime } from "@anilog/db/schema/anilog";
 import { user } from "@anilog/db/schema/auth";
-import { eq, getTableColumns, count, and, ilike, asc, or } from "drizzle-orm";
+import { eq, getTableColumns, count, and, ilike, asc } from "drizzle-orm";
 import { conflictError, internalError, notFoundError, validationError } from "../errors/api-error";
 
 export class UserService {
@@ -339,97 +339,11 @@ export class UserService {
     );
   }
 
-  static async getAdminStats() {
-    const [result] = await db.select({ count: count() }).from(user);
-    return { totalUsers: result?.count ?? 0 };
-  }
-
   static async getAdminStatus(userId: string) {
     const found = await db.query.user.findFirst({
       where: eq(user.id, userId),
       columns: { isAdmin: true },
     });
     return found?.isAdmin ?? false;
-  }
-
-  static async searchUsersForAdmin(
-    query: string,
-    options: { limit?: number; offset?: number } = {},
-  ) {
-    const limit = Math.min(Math.max(options.limit ?? 20, 1), 100);
-    const offset = Math.max(options.offset ?? 0, 0);
-    const normalized = query.trim();
-
-    const whereClause = normalized
-      ? or(
-          ilike(user.name, `%${normalized}%`),
-          ilike(user.username, `%${normalized}%`),
-          ilike(user.email, `%${normalized}%`),
-        )
-      : undefined;
-
-    const rows = await db
-      .select({
-        id: user.id,
-        name: user.name,
-        username: user.username,
-        email: user.email,
-        isAdmin: user.isAdmin,
-        image: user.image,
-        profile: getTableColumns(userProfile),
-      })
-      .from(user)
-      .leftJoin(userProfile, eq(user.id, userProfile.userId))
-      .where(whereClause)
-      .orderBy(asc(user.createdAt))
-      .limit(limit)
-      .offset(offset);
-
-    const [totalResult] = await db.select({ count: count() }).from(user).where(whereClause);
-
-    const users = await Promise.all(
-      rows.map(async (entry) => ({
-        id: entry.id,
-        name: entry.name,
-        username: entry.username,
-        email: entry.email,
-        isAdmin: entry.isAdmin,
-        image: entry.image,
-        profile: entry.profile,
-        followerCount: await this.getFollowerCount(entry.id),
-        followingCount: await this.getFollowingCount(entry.id),
-      })),
-    );
-
-    return {
-      users,
-      total: totalResult?.count ?? 0,
-      limit,
-      offset,
-    };
-  }
-
-  static async setUserAdminStatus(targetUserId: string, isAdmin: boolean, actorUserId?: string) {
-    if (actorUserId && actorUserId === targetUserId && !isAdmin) {
-      throw validationError("You cannot remove your own admin access");
-    }
-
-    const [updated] = await db
-      .update(user)
-      .set({
-        isAdmin,
-        updatedAt: new Date(),
-      })
-      .where(eq(user.id, targetUserId))
-      .returning({
-        id: user.id,
-        isAdmin: user.isAdmin,
-      });
-
-    if (!updated) {
-      throw notFoundError("User not found");
-    }
-
-    return updated;
   }
 }
